rename id67* to v1*
