+++ keyboards/idobao/id67/v1/v1.c
@@ -14,4 +14,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "id67.h"
+#include "v1.h"
